+++ distros/fedora/packages.sh
@@ -77,6 +77,7 @@ install_core_packages() {
         htop
         btop
         zoxide
+        cava
 
         # Theming & Appearance
         bibata-cursor-theme
@@ -104,6 +105,9 @@ install_core_packages() {
     )
 
     pkg_install "${core_pkgs[@]}"
+
+    # Verify Rustup / Cargo toolchain
+    check_rustup_cargo
 }
 
 install_greeter_packages() {